+++ BackEnd/auth/app/api/models/user_login_model.py
@@ -1,6 +1,5 @@
 from pydantic import BaseModel
 from typing import List, Optional
-from sqlalchemy import (Column, Integer, MetaData, String, Table, Float, create_engine, ARRAY)
